refactor code structures
