+++ openapi/v3_test.go
@@ -45,13 +45,12 @@ func TestOpenAPIv3ToRESTSchema(t *testing.T) {
 			var expected schema.NDCRestSchema
 			assertNoError(t, json.Unmarshal(expectedBytes, &expected))
 
-			output, errs := OpenAPIv3ToNDCSchema(sourceBytes, &ConvertOptions{
+			output, errs := OpenAPIv3ToNDCSchema(sourceBytes, ConvertOptions{
 				EnvPrefix:  tc.EnvPrefix,
 				TrimPrefix: "/v1",
 			})
 			if output == nil {
-				t.Error(errors.Join(errs...))
-				t.FailNow()
+				t.Fatal(errors.Join(errs...))
 			}
 
 			assertRESTSchemaEqual(t, &expected, output)
@@ -59,7 +58,7 @@ func TestOpenAPIv3ToRESTSchema(t *testing.T) {
 	}
 
 	t.Run("failure_empty", func(t *testing.T) {
-		_, err := OpenAPIv3ToNDCSchema([]byte(""), nil)
+		_, err := OpenAPIv3ToNDCSchema([]byte(""), ConvertOptions{})
 		assertError(t, errors.Join(err...), "there is nothing in the spec, it's empty")
 	})
 }
